+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 Nothing yet.
 
+## 2.12.1 - Thu Sep 17 2015
+
+### Fixes
+
+* AllMatchingTypesList - treats `stdClass` as a Traversable
+
 ## 2.12.0 - Wed Sep 16 2015
 
 ### Deprecated